Create file_event_win_remote_cred_dump.yml
After some testing, @ecote7 and I found out that when you use CrackMapExec or Impacket-secretsdump to dump LSA secrets against a remote endpoint, they'll be temporarily stored in a randomly named 8 characters .tmp file in the C:\Windows\System32 directory. Seems to be the case for other creds that gets dumped through this tool, even SAM.

Run the query against a few EDR endpoints and it seems to be high fidelity as a svchost.exe RemoteRegistry process will rarely create files there, if ever.

This is my 1st git push request ever, and my 1st time writing a Sigma rule. So if it can be adjusted, tweaked, fixed, etc. feel entirely free to do so and I'll appreciate any pointers you may have!