Update win_security_exploit_cve_2023_23397_outlook_remote_file_query.yml
'Accesses' field values may appear as code. The string value appears in EventViewer, but different SIEMs can store the original value as it is stored in xml (e.g. in Microsoft Sentinel this field is not extracted by default data connector and user should extract it during query time to be able to use it in detection). Also the field name in xml is AccessList but not Accesses.
https://learn.microsoft.com/en-us/windows/security/threat-protection/auditing/event-4656