Create net_connection_win_google_api_non_browser_access.yml
Created a rule to detect possible command and control activity associated with offensive tools utilizing the googleapis.com endpoints. In particular the use of the new GC2 framework that has been leveraged by current threat actors such as APT41.

Tested in uncoder.io and the resulting kusto query language was used in Microsoft Defender for Endpoint  (MDE) which detected the activity.