Modified rule to widen scope
Scope for this rule is too narrow. By limiting the parent processes, you're missing out on important chrome browser extensions. For example, Silver's CursedChrome would pass right through as would BurpSuite's chrome extension.

Instead the ParentImage is moved to the `fields` as it is interesting and important for the analyst, but should not narrow the scope of the search.