Create proc_creation_win_vuln_anydesk_detection.yml
This use case is designed to detect the presence and usage of the AnyDesk application with version less than 8.0.8 in your environment. AnyDesk is a remote desktop software that allows users to connect to computers remotely. While it has legitimate uses, it can also be exploited for malicious activities such as unauthorized access or data exfiltration.