Added Initiated field
I think this is required for sysmon logs but I could be wrong. This rule should work without this field as the destination port is 3389 but it matches form with other rules using the network_connection field.

It also helps my personal non-sysmon log backend to differentiate between network connect and network receive logs.