sensitive environment changes first pass (#46)
* Replace user_compromise with persistence

* Sensitive environment changes pass

* Updating to passthrough: aws_ec2_disable_encryption.yml

* Update to passthrough: azure_ad_user_added_to_admin_role.yml

* Update to passthrough: azure_firewall_modified_or_deleted.yml

* Update to passthrough: azure_firewall_rule_collection_modified_or_deleted.yml

* Disable rule.

Disabling due to conditional access policies overriding. Also o365 azure ad logs have been turned off in favor of azure dataset.

* Disable: aws_iam_backdoor_users_keys.yml

Broken logic in "filter" but valuable rule if we can fix

* Reverting tags: aws_iam_backdoor_users_keys.yml

Reverting tags as the issue is only on the filter to narrow down to "different user". This rule fires very rarely without that so we can re enable. If its a problem in the future we can work on converting this to EQL to fix the filter.

* Update to passthrough: azure_ad_new_root_ca_added.yml

* Update to passthrough: azure_ad_users_added_to_device_admin_roles.yml

* Remove tags from container registry

---------

Co-authored-by: peterydzynski <25185548+peterydzynski@users.noreply.github.com>