Fix proc_creation_win_filefix_browsers.yml
`#` does not appear in command line logs a  when a filefix attack is executed., at least not nin Falcon FDR logs. Instead, the executable is always wrapped in double quotes. These quotes are not present when extensions are executed.

This PR updates the CommandLine to reflect this  using regex, while maintaining the original CommandLine detection of `#` in case that is still used in other logs, and adds more browser image names