Update proc_creation_win_werfaultsecure_process_freeze.yml
Fixed to explicitly exclude werfaultsecure.exe when running out of known trusted/expected file path's. Referenced write up states "We can prevent the exploitation of this vulnerability by monitoring the image file location of the WerFaultSecure.exe process. If it is located outside of System32, there is a high likelihood that it is malicious."